Start on support for ints > 32-bits.
